deduplicate and simplify the manifest reading logic
