+++ src/include/iceberg_types.hpp
@@ -16,56 +16,7 @@
 
 namespace duckdb {
 
-//! fwd declare
-struct ManifestReaderInput;
-
 using sequence_number_t = int64_t;
-
-enum class IcebergManifestContentType : uint8_t {
-	DATA = 0,
-	DELETE = 1,
-};
-
-static string IcebergManifestContentTypeToString(IcebergManifestContentType type) {
-	switch (type) {
-	case IcebergManifestContentType::DATA:
-		return "DATA";
-	case IcebergManifestContentType::DELETE:
-		return "DELETE";
-	default:
-		throw InvalidConfigurationException("Invalid Manifest Content Type");
-	}
-}
-
-enum class IcebergManifestEntryStatusType : uint8_t { EXISTING = 0, ADDED = 1, DELETED = 2 };
-
-static string IcebergManifestEntryStatusTypeToString(IcebergManifestEntryStatusType type) {
-	switch (type) {
-	case IcebergManifestEntryStatusType::EXISTING:
-		return "EXISTING";
-	case IcebergManifestEntryStatusType::ADDED:
-		return "ADDED";
-	case IcebergManifestEntryStatusType::DELETED:
-		return "DELETED";
-	default:
-		throw InvalidConfigurationException("Invalid matifest entry type");
-	}
-}
-
-enum class IcebergManifestEntryContentType : uint8_t { DATA = 0, POSITION_DELETES = 1, EQUALITY_DELETES = 2 };
-
-static string IcebergManifestEntryContentTypeToString(IcebergManifestEntryContentType type) {
-	switch (type) {
-	case IcebergManifestEntryContentType::DATA:
-		return "EXISTING";
-	case IcebergManifestEntryContentType::POSITION_DELETES:
-		return "POSITION_DELETES";
-	case IcebergManifestEntryContentType::EQUALITY_DELETES:
-		return "EQUALITY_DELETES";
-	default:
-		throw InvalidConfigurationException("Invalid Manifest Entry Content Type");
-	}
-}
 
 struct FieldSummary {
 	bool contains_null = false;
@@ -75,6 +26,11 @@ struct FieldSummary {
 	Value lower_bound;
 	//! Optional
 	Value upper_bound;
+};
+
+enum class IcebergManifestContentType : uint8_t {
+	DATA = 0,
+	DELETE = 1,
 };
 
 //! An entry in the manifest list file (top level AVRO file)
@@ -96,13 +52,6 @@ public:
 	vector<FieldSummary> field_summary;
 
 public:
-	static idx_t ProduceEntries(DataChunk &chunk, idx_t offset, idx_t count, const ManifestReaderInput &input,
-	                            vector<entry_type> &entries);
-	static bool VerifySchema(const case_insensitive_map_t<ColumnIndex> &name_to_vec);
-	static void PopulateNameMapping(idx_t column_id, const LogicalType &type, const string &name,
-	                                case_insensitive_map_t<ColumnIndex> &name_to_vec);
-
-public:
 	static vector<LogicalType> Types() {
 		return {
 		    LogicalType::VARCHAR,
@@ -111,10 +60,25 @@ public:
 		};
 	}
 
+	static string ContentTypeToString(IcebergManifestContentType type) {
+		switch (type) {
+		case IcebergManifestContentType::DATA:
+			return "DATA";
+		case IcebergManifestContentType::DELETE:
+			return "DELETE";
+		default:
+			throw InvalidConfigurationException("Invalid Manifest Content Type");
+		}
+	}
+
 	static vector<string> Names() {
 		return {"manifest_path", "manifest_sequence_number", "manifest_content"};
 	}
 };
+
+enum class IcebergManifestEntryContentType : uint8_t { DATA = 0, POSITION_DELETES = 1, EQUALITY_DELETES = 2 };
+
+enum class IcebergManifestEntryStatusType : uint8_t { EXISTING = 0, ADDED = 1, DELETED = 2 };
 
 //! An entry in a manifest file
 struct IcebergManifestEntry {
@@ -140,17 +104,36 @@ public:
 	int64_t file_size_in_bytes;
 
 public:
-	static idx_t ProduceEntries(DataChunk &chunk, idx_t offset, idx_t count, const ManifestReaderInput &input,
-	                            vector<entry_type> &entries);
-	static bool VerifySchema(const case_insensitive_map_t<ColumnIndex> &name_to_vec);
-	static void PopulateNameMapping(idx_t column_id, const LogicalType &type, const string &name,
-	                                case_insensitive_map_t<ColumnIndex> &name_to_vec);
-
-public:
 	static vector<LogicalType> Types() {
 		return {
 		    LogicalType::VARCHAR, LogicalType::VARCHAR, LogicalType::VARCHAR, LogicalType::VARCHAR, LogicalType::BIGINT,
 		};
+	}
+
+	static string ContentTypeToString(IcebergManifestEntryContentType type) {
+		switch (type) {
+		case IcebergManifestEntryContentType::DATA:
+			return "EXISTING";
+		case IcebergManifestEntryContentType::POSITION_DELETES:
+			return "POSITION_DELETES";
+		case IcebergManifestEntryContentType::EQUALITY_DELETES:
+			return "EQUALITY_DELETES";
+		default:
+			throw InvalidConfigurationException("Invalid Manifest Entry Content Type");
+		}
+	}
+
+	static string StatusTypeToString(IcebergManifestEntryStatusType type) {
+		switch (type) {
+		case IcebergManifestEntryStatusType::EXISTING:
+			return "EXISTING";
+		case IcebergManifestEntryStatusType::ADDED:
+			return "ADDED";
+		case IcebergManifestEntryStatusType::DELETED:
+			return "DELETED";
+		default:
+			throw InvalidConfigurationException("Invalid matifest entry type");
+		}
 	}
 
 	static vector<string> Names() {
